back-ported DBCommandCursor from upstream #503
This is a cherry-pick of about a third of
mongodb/mongo@a51f2688fa05672d999c997170847a3ee29a223b, ignoring the
aggregation pieces because they are more complexity than we want to take
from a pre-release version of upstream.

Commands that use this will either need a fall-back mechanism or will need
explicit driver support that is coming in upstream 2.6.  For now it's
probably ok to just make sure there are js shell wrappers for anything we
use it for (see the remainder of the change linked above) and assume
they're admin-only-ish commands until driver support becomes widespread.

closes #503